requirements.txt
# requirements.txt

flask
transformers
torch
langdetect
pickle5
pandas
sklearn
google-generativeai